requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.13.0
opencv-python>=4.8.0
Pillow>=9.4.0
tqdm>=4.65.0
streamlit>=1.30.0
wandb>=0.16.0
grad-cam==1.5.5

# pip install -r requirements.txt